+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-03 13:12:43
-// Commit: feat(core): implement responsive core view component
+// Updated at: 2026-06-03 13:39:40
+// Commit: chore(core): update dependencies and cleanup unused imports in core
 
 class coreComponentState {
-  final String commitRef = '1693';
-  final String lastMessage = 'feat(core): implement responsive core view component';
-  final DateTime timestamp = DateTime.parse('2026-06-03T13:12:43+05:00');
+  final String commitRef = '1698';
+  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
+  final DateTime timestamp = DateTime.parse('2026-06-03T13:39:40+05:00');
 }